cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pkm"
version = "0.2.0"
description = "Personal Knowledge Management CLI for Obsidian vaults"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.scripts]
pkm = "pkm.cli:main"

[project.optional-dependencies]
search = [
    "sentence-transformers>=2.0",
    "numpy>=1.24",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/pkm"]

[tool.pytest.ini_options]
testpaths = ["tests"]
